refactor(decompose): split packages/libraries/graph-tools/src/commands/session/nodeClick.ts into folder
